adjusting seed files + preparing for publishing + finishing pagination

+++ Core/Entites/Benchmark/GPUBenchmark.cs
@@ -13,7 +13,7 @@ namespace Core.Entites.Benchmark
         public int? G3DMark { get; set; } // Graphics 3d Mark
         public int? G2DMark { get; set; } // Graphics 2D Mark
         public double? GpuValue { get; set; }
-        public int? TDP { get; set; }
+        public double? TDP { get; set; }
         public double? PowerPerformance { get; set; }
         public int? TestDate { get; set; }
         public string? Category { get; set; }
